requirements.txt
numpy>=1.20.0
scipy>=1.7.0
qutip>=4.6.0
matplotlib>=3.4.0
pytest
black
mypy
pylint
tqdm>=4.62.0
ipykernel
jupyter
numba>=0.54.0
cupy-cuda11x>=11.0.0  # For CUDA 11.x, adjust version based on your CUDA installation
